Seperate inner class on MapsViewFragment.kt to Class CustomMarker
